Issue 537: Take namespaces that are configured to be ignored into account
Although one can provide an array of namespaces that shall be ignored to createClient(), they are not used in objectToXml(). This commit fixes that.

Moreover it removes the shouldIgnoreNamespace() method because it makes no sense to explicitly configure namespaces to be ignored and than let some weird logic calculate if a given namespace is going to be ignored.